Fix SLSQP package
The package previously did not have an init file. Versions of setuptools older than 3.0 would automatically add init file to the root package if it was missing. This was removed from setuptools-3.0 and forward. Without the init file, the package cannot be imported. This fix should allow creating a distritbuion that imports correctly regardless of the version of setuptools that was used.