Fix package setup for COBYLA
Applied a similar fix that was used for SLSQP. This fix should ensure that once
the distribution is installed, it can be imported regardless of the version
of setuptools that was used when the distribution was created.

The same fix should be applied to CONMIN, NEWTSUM and any remaining
distributions we build and maintain are the intended to be importable Python
packages.